% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypic_values.R
\name{genotypic_values}
\alias{genotypic_values}
\title{Initialize genotypic value of offspring from ploidy with respect to the focal allele}
\usage{
genotypic_values(allele_names, ploidy)
}
\description{
Initialize bi-allele multi-locus genotype frequency from underlying allele frequencies assuming no prior linkage disequilibrium
}
\details{
@param allele_names A vector of designated letters associated with each locus, with upper-case denoting the focal allele and lower-case denoting the non-focal (e.g. A vs a)
 @param ploidy A vector of the ploidies of each locus

 @return A vector of the genotype frequencies
 @export
 @examples
}
