% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_femalefitness.R
\name{initialize_femalefitness}
\alias{initialize_femalefitness}
\title{Initialize female fitness of different genotypes}
\usage{
initialize_femalefitness(
  pk,
  zk,
  in_use,
  insecticides_applied,
  resistant_alleles,
  strategy,
  strategy_details,
  ins_dec,
  mixmax
)
}
\description{
Initialize fitness of bi-allele multi-locus genotypes of females from underlying selection coefficients frequencies
}
\details{
@param pk The constant for the percentage of the human population protected by the insecticide
 @param tk The constant for the toxicity cost of the insecticide e.g. mosquitio mortality
 @param ak The constant for the repellency of the insecticide as a (negative) assortment factor
 @param resistant_alleles A matrix that encodes the properties of the resistant alleles, as one allele per row
 The resistant alleles matrix takes the form of:
 name (i.e. a letter), 1=haploid or 2=diploid, T=toxicity or A=assortment, effect, dominance of effect, cost, dominance of cost

 @return A vector of the fitness of different genotypes
 @export
 @examples
}
