% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_freqset.R
\name{initialize_freqset}
\alias{initialize_freqset}
\title{Initialize genotype frequency from allele frequency}
\usage{
initialize_freqset(allele_frequencies, allele_names, ploidy)
}
\description{
Initialize bi-allele multi-locus genotype frequency from underlying allele frequencies assuming no prior linkage disequilibrium
}
\details{
@param allele_frequencies A vector of frequencies 'f' of the focal alleles (upper-case, e.g. A) across all loci, where non-focal alleles (lower-case, e.g. a) have frequency 1-f
 @param allele_names A vector of designated letters associated with each locus, with upper-case denoting the focal allele and lower-case denoting the non-focal (e.g. A vs a)
 @param ploidy A vector of the ploidies of each locus

 @return A vector of the genotype frequencies
 @export
 @examples
}
