% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_malefitness.R
\name{initialize_malefitness}
\alias{initialize_malefitness}
\title{Initialize male fitness of different genotypes}
\usage{
initialize_malefitness(wfset, ek, resistant_alleles)
}
\description{
Initialize fitness of bi-allele multi-locus genotypes of males from underlying selection coefficients frequencies, assuming that a male has percentage similar fitness to a female but always pays the cost of resistance
}
\details{
@param wfset The vector of female fitness
 @param ek The constant for the degree to which males and females experience the same fitness
 @param resistant_alleles A matrix that encodes the properties of the resistant alleles, as one allele per row
 The resistant alleles matrix takes the form of:
 name (i.e. a letter), 1=haploid or 2=diploid, T=toxicity or A=assortment, effect, dominance of effect, cost, dominance of cost

 @return A vector of the fitness of different genotypes
 @export
 @examples
}
