% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mating.R
\name{random_mating}
\alias{random_mating}
\title{Sexual recombination that occurs freely with potential linkage disequilibrium}
\usage{
random_mating(ffset, fmset, resistant_alleles)
}
\description{
Provides offspring bi-allele multi-locus frequencies under the assumption of panmixia, where female / male fitness can be different and selection can build linkage disequilibrium
}
\details{
@param ffset A vector with the frequency of all female genotypes
 @param fmset A vector with the frequency of all male genotypes
 @param resistant_alleles A matrix that encodes the properties of the resistant alleles, as one allele per row

 @return A vector with the frequency of all genotypes
 @export
 @examples
}
