% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinitialize_freqset.R
\name{reinitialize_freqset}
\alias{reinitialize_freqset}
\title{Reinitialize genotype frequency from allele frequency}
\usage{
reinitialize_freqset(aset_added, fset, aset_add, anames, ploidy, add_loci)
}
\description{
Reinitialize bi-allele multi-locus genotype frequency from underlying allele frequencies assuming no prior linkage disequilibrium
}
\details{
@param aset_added A vector with the frequency of all alleles
 @param fset A vector with the frequency of all genotypes
 @param aset_add A vector of frequencies 'f' of the focal alleles (upper-case, e.g. A) across all loci, where non-focal alleles (lower-case, e.g. a) have frequency 1-f
 @param anames A vector of designated letters associated with each locus, with upper-case denoting the focal allele and lower-case denoting the non-focal (e.g. A vs a)
 @param ploidy A vector of the ploidies of each locus
 @param add_loci A vector of the number of loci to be added

 @return A vector of the genotype frequencies
 @export
 @examples
}
