% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_strategy.R
\name{update_strategy}
\alias{update_strategy}
\title{Update insecticides that are in use in line with the strategy}
\usage{
update_strategy(
  tt,
  strategy,
  strategy_start,
  strategy_details,
  res_freq,
  in_use,
  rk
)
}
\description{
Expresses strategy as a change in insecticide usage dependent upon conditions for when a strategy involves changing an insecticide
}
\details{
@param tt Current time t in generations
 @param strategy Either Combination, Mixture or Rotation
 @param strategy_start The time t in generations that the strategy starts
 @param strategy_details A vector informing the strategy
 @param res_freq The frequency of resistant alleles within the populaion
 @param in_use The previous vector of in_use
 @param rk THe replacement / retreatment schedule

 @return A vector of the fitness of different genotypes
 @export
 @examples
}
